requirements.txt
openai>=1.0.0
azure-communication-email>=1.0.0
azure-functions
